Label
label elements are used to help associate the text for an input element with the input element itself (especially for assistive technologies like screen readers). For example, <label><input type="radio"> Indoor</label> makes it so clicking the word cat also selects the corresponding radio button.
It should have a space between the radio and Indoor